+++ src/Field/Choices/DatasourceChoices.php
@@ -31,12 +31,10 @@ final class DatasourceChoices implements ChoicesInterface
 	/**
 	 * @inheritDoc
 	 */
-	public function isValidData (
-		array|int|string $data,
-		?ComponentContext $context = null,
-	) : bool
+	public function getValidationConstraints (bool $allowMultiple) : array
 	{
-		return true;
+		// always valid
+		return [];
 	}
 
 	/**