+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * (bc) Remove ability to filter by component groups.
 * (improvement) Add internal `LocaleHelper` to unify locale detection.
 * (feature) Add `StoryMetaData::getTranslatedDocumentsMapping()` to fetch the required data to map hreflangs.
+* (deprecation) Deprecate `AbstractField::enablePreview()` in favor of `AbstractField::useAsAdminDisplayName()`.
 
 
 3.0.0-beta.5